+++ packages/embed-wallet/src/constants.ts
@@ -1,1 +1,1 @@
-export const WALLET_CLIENT_VERSION = 'v1.24.0-rc.1';
+export const WALLET_CLIENT_VERSION = 'v1.24.0-rc.2';